+++ src/services/credentials.ts
@@ -4,7 +4,7 @@ import {
   VC_CONTEXT,
   VC_TYPE
 } from '../types/constants.js'
-import { CredentialRequest, VerifyStatusOptions } from '../types/types.js'
+import { CredentialRequest, VerificationOptions } from '../types/types.js'
 import { Identity } from './identity/index.js'
 import { VeridaService } from '../services/connectors/verida.js'
 import { v4 } from 'uuid'
@@ -50,15 +50,4 @@ export class Credentials {
         }
         return verifiable_credential
 	}
-
-	async verify_credentials(credential: W3CVerifiableCredential | string,  statusOptions: VerifyStatusOptions | null, agentId: string): Promise<IVerifyResult> {
-		const result = await Identity.instance.verifyCredential(credential, statusOptions, agentId)
-        delete(result.payload)
-        return result
-	}
-
-    async verify_presentation(presentation: W3CVerifiablePresentation, agentId: string): Promise<IVerifyResult> {
-        const result = await Identity.instance.verifyPresentation(presentation, agentId)
-        return result
-    }
 }