+++ src/services/identity/agent.ts
@@ -14,26 +14,32 @@ import {
   MinimalImportableKey,
   TAgent,
   VerifiableCredential,
-  VerifiablePresentation,
+  VerifiablePresentation
 } from '@veramo/core'
 import { KeyManager } from '@veramo/key-manager'
 import { DIDStore, KeyStore } from '@veramo/data-store'
 import { DIDManager } from '@veramo/did-manager'
-import { DIDResolverPlugin } from '@veramo/did-resolver'
+import { DIDResolverPlugin, getUniversalResolver as UniversalResolver } from '@veramo/did-resolver'
+import { getResolver as VeridaResolver } from '@verida/vda-did-resolver'
 import { CredentialPlugin } from '@veramo/credential-w3c'
 import { CredentialIssuerLD, LdDefaultContexts, VeramoEd25519Signature2018 } from '@veramo/credential-ld'
 import { Cheqd, getResolver as CheqdDidResolver, ResourcePayload } from '@cheqd/did-provider-cheqd'
+import { getDidKeyResolver as KeyDidResolver } from '@veramo/did-provider-key'
 import { CheqdNetwork } from '@cheqd/sdk'
 import { Resolver, ResolverRegistry } from 'did-resolver'
-import { fromString } from 'uint8arrays'
-import {
+import type {
+  ICheqdBroadcastStatusList2021Args,
+  ICheqdCheckCredentialStatusWithStatusList2021Args,
   ICheqdCreateStatusList2021Args,
-  ICheqdGenerateStatusList2021Args,
-  ICheqdVerifyCredentialWithStatusList2021Args
-} from '@cheqd/did-provider-cheqd/build/types/agent/ICheqd.js'
-import { v4 } from 'uuid'
-
+  ICheqdDeactivateIdentifierArgs,
+  ICheqdRevokeBulkCredentialsWithStatusList2021Args,
+  ICheqdUpdateIdentifierArgs,
+  ICheqdVerifyCredentialWithStatusList2021Args,
+  ICheqdVerifyPresentationWithStatusList2021Args,
+} from '@cheqd/did-provider-cheqd/build/types/agent/ICheqd'
 import {
+  BroadCastStatusListOptions,
+  CheckStatusListOptions,
   cheqdDidRegex,
   CreateAgentRequest,
   CreateStatusListOptions,
@@ -41,10 +47,12 @@ import {
   RevocationStatusOptions,
   StatusOptions,
   SuspensionStatusOptions,
+  UpdateStatusListOptions,
   VeramoAgent,
-  VerifyStatusOptions
+  VerificationOptions
 } from '../../types/types.js'
 import { VC_PROOF_FORMAT, VC_REMOVE_ORIGINAL_FIELDS } from '../../types/constants.js'
+import { decodeJWT } from 'did-jwt'
 
 const resolverUrl = "https://resolver.cheqd.net/1.0/identifiers/"
 
@@ -80,7 +88,10 @@ export class Veramo {
         plugins.push(
           new DIDResolverPlugin({
             resolver: new Resolver({
-              ...CheqdDidResolver({ url: process.env.RESOLVER_URL }) as ResolverRegistry
+              ...CheqdDidResolver({ url: process.env.RESOLVER_URL }) as ResolverRegistry,
+              ...KeyDidResolver(),
+              ...VeridaResolver(),
+              ...UniversalResolver()
             })
           })
         )
@@ -123,6 +134,38 @@ export class Veramo {
         }
       })
       return identifier
+    } catch (error) {
+      throw new Error(`${error}`)
+    }
+  }
+
+  async updateDid(agent: VeramoAgent, didDocument: DIDDocument): Promise<IIdentifier> {
+    try {
+      const [kms] = await agent.keyManagerGetKeyManagementSystems()
+
+      const result = await agent.cheqdUpdateIdentifier({
+        kms,
+        document: didDocument,
+      } satisfies ICheqdUpdateIdentifierArgs)
+      return {...result, provider: 'cheqd'}
+    } catch (error) {
+      throw new Error(`${error}`)
+    }
+  }
+
+  async deactivateDid(agent: VeramoAgent, did: string): Promise<boolean> {
+    try {
+      const [kms] = await agent.keyManagerGetKeyManagementSystems()
+      const didDocument = (await this.resolveDid(agent, did)).didDocument
+
+      if (!didDocument) {
+        throw new Error('DID document not found')
+      }
+      const result = await agent.cheqdDeactivateIdentifier({
+        kms,
+        document: didDocument
+      } satisfies ICheqdDeactivateIdentifierArgs)
+      return result
     } catch (error) {
       throw new Error(`${error}`)
     }
@@ -196,46 +239,117 @@ export class Veramo {
     }          
   }
 
-  async verifyCredential(agent: VeramoAgent, credential: string | VerifiableCredential, statusOptions: VerifyStatusOptions | null): Promise<IVerifyResult> {
-    if(typeof credential !== 'string') {
-        return await agent.cheqdVerifyCredential({
+  async verifyCredential(agent: VeramoAgent, credential: string | VerifiableCredential, verificationOptions: VerificationOptions = {}): Promise<IVerifyResult> {
+    const decodedCredential = typeof credential === 'string' ? decodeJWT(credential) as unknown as VerifiableCredential : credential
+    let result: IVerifyResult
+    if(verificationOptions.verifyStatus) {
+        result = await agent.cheqdVerifyCredential({
             credential: credential as VerifiableCredential,
             fetchList: true,
-            ...statusOptions
+            verificationArgs: {
+                ...verificationOptions,
+                fetchRemoteContexts: verificationOptions.fetchRemoteContexts || decodedCredential.proof.jws
+            }
         } as ICheqdVerifyCredentialWithStatusList2021Args)
+    } else {
+        result = await agent.verifyCredential({
+            credential, 
+            ...verificationOptions,
+            fetchRemoteContexts: verificationOptions.fetchRemoteContexts || decodedCredential.proof.jws
+        })
     }
-    return await agent.verifyCredential({ credential, fetchRemoteContexts: true })
+
+    if (result.didResolutionResult) {
+        delete(result.didResolutionResult)
+    }
+
+    if (result.jwt) {
+        delete(result.jwt)
+    }
+
+    if (result.verifiableCredential) {
+        delete(result.verifiableCredential)
+    }
+
+    if (result.payload) {
+        delete(result.payload)
+    }
+
+    return result
   }
 
-  async verifyPresentation(agent: VeramoAgent, presentation: VerifiablePresentation | string): Promise<IVerifyResult> {
-    return await agent.verifyPresentation({ presentation, fetchRemoteContexts: true })
+  async verifyPresentation(agent: VeramoAgent, presentation: VerifiablePresentation | string, verificationOptions: VerificationOptions = {}): Promise<IVerifyResult> {
+    let result: IVerifyResult
+    if(verificationOptions.verifyStatus) {
+        result = await agent.cheqdVerifyPresentation({
+            presentation: presentation as VerifiablePresentation,
+            fetchList: true,
+            verificationArgs: {
+                ...verificationOptions
+            },
+        } as ICheqdVerifyPresentationWithStatusList2021Args)
+    } else {
+        result = await agent.verifyPresentation({ presentation, ...verificationOptions, fetchRemoteContexts: verificationOptions.fetchRemoteContexts || false })
+    }
+
+    if (result.didResolutionResult) {
+        delete(result.didResolutionResult)
+    }
+
+    if (result.jwt) {
+        delete(result.jwt)
+    }
+
+    if (result.verifiablePresentation) {
+        delete(result.verifiablePresentation)
+    }
+
+    if (result.payload) {
+        delete(result.payload)
+    }
+
+    return result
   }
 
-  async createStatusList2021(agent: VeramoAgent, did: string, network: string, resourceOptions: ResourcePayload, statusOptions: CreateStatusListOptions) {
-    const statusList = await agent.cheqdGenerateStatusList2021({
-      buffer: resourceOptions.data,
-      length: statusOptions.length,
-      bitstringEncoding: statusOptions.encoding
-    } as ICheqdGenerateStatusList2021Args)
-
+  async createStatusList2021(agent: VeramoAgent, did: string, resourceOptions: ResourcePayload, statusOptions: CreateStatusListOptions) {
     const [kms] = await agent.keyManagerGetKeyManagementSystems()
 
+    if (!resourceOptions.name) {
+        throw new Error(`StatusList name is required`)
+    }
     return await agent.cheqdCreateStatusList2021({
       kms,
-      payload: {
-         collectionId: did.split(':')[3],
-         data: fromString(statusList, statusOptions.encoding || 'base64url'),
-         resourceType: 'StatusList2021',
-         version: resourceOptions.version,
-         name: resourceOptions.name,
-         id: v4(),
-      },
-      network: network as CheqdNetwork
-   } as ICheqdCreateStatusList2021Args)
+      issuerDid: did,
+      statusListName: resourceOptions.name,
+      statusPurpose: statusOptions.statusPurpose || 'revocation',
+      statusListEncoding: statusOptions.encoding || 'base64url',
+      statusListLength: statusOptions.length,
+      encrypted: statusOptions.encrypted || false,
+      resourceVersion: resourceOptions.version
+   } satisfies ICheqdCreateStatusList2021Args)
+ }
+
+ async broadcastStatusList2021(agent: VeramoAgent, did: string, resourceOptions: ResourcePayload, statusOptions: BroadCastStatusListOptions) {
+    const [kms] = await agent.keyManagerGetKeyManagementSystems()
+
+    if (!resourceOptions.data) {
+        throw new Error(`StatusList data is required`)
+    }
+
+    return await agent.cheqdBroadcastStatusList2021({
+        kms,
+        payload: {
+            ...resourceOptions,
+            collectionId: did.split(':')[3],
+            data: resourceOptions.data,
+            resourceType: statusOptions.statusPurpose === 'revocation' ? 'StatusList2021Revocation' : 'StatusList2021Suspension'
+        },
+        network: did.split(':')[2] as CheqdNetwork,
+     } satisfies ICheqdBroadcastStatusList2021Args)
  }
 
  async revokeCredentials(agent: VeramoAgent, credentials: VerifiableCredential | VerifiableCredential[], publish: boolean=true) {
-    if (Array.isArray(credentials)) return await agent.cheqdRevokeCredentials({ credentials, fetchList: true, publish: true })
+    if (Array.isArray(credentials)) return await agent.cheqdRevokeCredentials({ credentials, fetchList: true, publish: true } satisfies ICheqdRevokeBulkCredentialsWithStatusList2021Args)
     return await agent.cheqdRevokeCredential({ credential: credentials, fetchList: true, publish })
  }
 
@@ -255,5 +369,56 @@ export class Veramo {
  async unsuspendCredentials(agent: VeramoAgent, credentials: VerifiableCredential | VerifiableCredential[], publish: boolean=true) {
     if (Array.isArray(credentials)) return await agent.cheqdUnsuspendCredentials({ credentials, fetchList: true, publish })
     return await agent.cheqdUnsuspendCredential({ credential: credentials, fetchList: true, publish })
+ }
+
+ async updateStatusList2021(agent: VeramoAgent, did: string, statusOptions: UpdateStatusListOptions, publish: boolean=true) {
+    switch(statusOptions.statusAction) {
+        case 'revoke': 
+            return await agent.cheqdRevokeCredentials({
+                revocationOptions: {
+                    issuerDid: did,
+                    statusListIndices: statusOptions.indices,
+                    statusListName: statusOptions.statusListName,
+                    statusListVersion: statusOptions.statusListVersion
+                },
+                fetchList: true,
+                publish,
+                returnUpdatedStatusList: !publish
+            })
+        case 'suspend':
+            return await agent.cheqdSuspendCredentials({
+                suspensionOptions: {
+                    issuerDid: did,
+                    statusListIndices: statusOptions.indices,
+                    statusListName: statusOptions.statusListName,
+                    statusListVersion: statusOptions.statusListVersion
+                },
+                fetchList: true,
+                publish,
+                returnUpdatedStatusList: !publish
+            })
+        case 'reinstate':
+            return await agent.cheqdUnsuspendCredentials({
+                unsuspensionOptions: {
+                    issuerDid: did,
+                    statusListIndices: statusOptions.indices,
+                    statusListName: statusOptions.statusListName,
+                    statusListVersion: statusOptions.statusListVersion
+                },
+                fetchList: true,
+                publish,
+                returnUpdatedStatusList: !publish
+            })           
+    }
+ }
+
+ async checkStatusList2021(agent: VeramoAgent, did: string, statusOptions: CheckStatusListOptions) {
+    return await agent.cheqdCheckCredentialStatus({
+        statusOptions: {
+            issuerDid: did,
+            ...statusOptions,
+        },
+        fetchList: true
+    } satisfies ICheqdCheckCredentialStatusWithStatusList2021Args)
  }
 }